Dr.Ambedkar-RAG/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
google-genai
python-dotenv
numpy

python-dotenv
requests
